api/requirements.txt
Django>=4.2,<6
djangorestframework>=3.14
django-cors-headers>=4.4
Pillow>=10.0
python-dotenv>=1.0
requests